Namespace renamed from Templating to Template
