+++ IntraWeb/test/IntraWeb.UnitTests/Service/Template/TemplateDataConverterTests.cs
@@ -1,7 +1,7 @@
-﻿using IntraWeb.Services.Templating;
+﻿using IntraWeb.Services.Template;
 using Xunit;
 
-namespace IntraWeb.UnitTests.Service.Templating
+namespace IntraWeb.UnitTests.Service.Template
 {
     public class TemplateDataConverterTests
     {
